Add exception when an 'ui' is expected, but not defined. Add activateItem() in menu to retrieve the actual route. Move template version to util.utils.py Fix BUG in backupFile() now test for file existance. rename 'case-insensitive-db-ids' to 'ignore-case-db-ids' to make both command line and options group in the template the same.

+++ gencrud/util/utils.py
@@ -28,6 +28,7 @@ backupFiles     = False
 overWriteFiles  = False
 lowerCaseDbIds  = False
 useModule       = False
+version         = 1
 
 C_FILEMODE_UPDATE = 'r+'
 C_FILEMODE_WRITE  = 'w'
@@ -42,7 +43,8 @@ def backupFile( file_name ):
         while os.path.isfile( file_name + '.~{0}'.format( idx ) ):
             idx += 1
 
-        shutil.copyfile( file_name, file_name + '.~{0}'.format( idx ) )
+        if os.path.isfile( file_name ):
+            shutil.copyfile( file_name, file_name + '.~{0}'.format( idx ) )
 
     return
